refactor(norm)!: restrict `token` definition
Component name, namespace and Composition name now cannot contain `-` or `_`
