+++ features/bridges/node/syntaxes.feature
@@ -4,7 +4,7 @@ Feature: Node algorithm implementation syntaxes
 
   Scenario Outline: Run <type> with <syntax> syntax
     Given I boot `node.syntaxes` component
-    When I invoke `<type>_<syntax>` with:
+    When I invoke `<type><syntax>` with:
       """yaml
       query: {}
       """
@@ -15,19 +15,19 @@ Feature: Node algorithm implementation syntaxes
       """
     Examples:
       | syntax   | type        |
-      | function | transition  |
-      | class    | transition  |
-      | factory  | transition  |
-      | function | observation |
-      | class    | observation |
-      | factory  | observation |
-      | function | assignment  |
-      | class    | assignment  |
-      | factory  | assignment  |
+      | Function | transition  |
+      | Class    | transition  |
+      | Factory  | transition  |
+      | Function | observation |
+      | Class    | observation |
+      | Factory  | observation |
+      | Function | assignment  |
+      | Class    | assignment  |
+      | Factory  | assignment  |
 
   Scenario Outline: Run <type> with <syntax> syntax
     Given I boot `node.syntaxes` component
-    When I invoke `<type>_<syntax>`
+    When I invoke `<type><syntax>`
     And I disconnect
     Then the reply is received:
       """
@@ -35,10 +35,9 @@ Feature: Node algorithm implementation syntaxes
       """
     Examples:
       | syntax   | type        |
-      | function | computation |
-      | class    | computation |
-      | factory  | computation |
-      | function | effect      |
-      | class    | effect      |
-      | factory  | effect      |
-
+      | Function | computation |
+      | Class    | computation |
+      | Factory  | computation |
+      | Function | effect      |
+      | Class    | effect      |
+      | Factory  | effect      |
