+++ features/steps/.workspace/components/collection/node.syntaxes/operations/assignmentClass.js
