+++ features/steps/.workspace/components/collection/node.syntaxes/operations/transitionClass.js
