+++ features/steps/.workspace/components/collection/node.syntaxes/operations/transitionFunction.js
