+++ features/steps/.workspace/components/collection/origins.amqpCredentials/manifest.toa.yaml
@@ -1,4 +1,4 @@
-name: amqp_credentials
+name: amqpCredentials
 namespace: origins
 
 origins:
