+++ features/steps/.workspace/components/collection/origins.httpAbsolute/manifest.toa.yaml
@@ -1,4 +1,4 @@
-name: http_absolute
+name: httpAbsolute
 namespace: origins
 
 operations:
