+++ features/steps/.workspace/components/collection/origins.httpEcho/manifest.toa.yaml
